Create Currency_Converter
An interactive program to convert currency. pulls data from aopen source website called float rates.com. First asks the user which currency theywant to convert from. Then asks what currency they want to convert to and how much. An infinite loop is used for continous interacation. A break occurs when no or a blank currency is implemented or a false currency is requested.